requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Data validation
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0  # Compatibility with passlib 1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Background tasks
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
